# StayLocal_OpenAPI.yaml
openapi: 3.0.3
info:
  title: StayLocal Backend API
  version: 1.0.0
  description: Contract-first REST API for the StayLocal community tourism platform.
    Covers authentication, destinations, properties, media, camera+GPS verification,
    bookings, rewards, reviews and admin moderation.
  contact:
    name: StayLocal Project Team
    email: staylocal-team@example.com
servers:
- url: http://localhost:8787
  description: Local development
- url: http://localhost:8787
  description: Demo/CI default
...
paths:
...
  /api/v1/uploads/init:
    post:
      operationId: initUpload
      summary: Create a signed media upload
      description: Returns a short-lived upload URL or an equivalent local-storage
        target. The API validates media type and size before issuing the upload contract.
      tags:
      - Media
      security: *id003
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - filename
              - contentType
              - sizeBytes
              properties:
                filename:
                  type: string
                  minLength: 1
                  maxLength: 255
                contentType:
                  type: string
                  enum:
                  - image/jpeg
                  - image/png
                  - image/webp
                sizeBytes:
                  type: integer
                  minimum: 1
                  maximum: 10485760
            examples:
              201_upload_init_success:
                value:
                  filename: room.jpg
                  contentType: image/jpeg
                  sizeBytes: 524288
              422_upload_too_large:
                value:
                  filename: room.jpg
                  contentType: image/jpeg
                  sizeBytes: 20000000
      responses:
        '201':
          description: Upload initialized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadInit'
              examples:
                201_upload_init_success:
                  value:
                    assetId: asset-1001
                    uploadUrl: https://storage.example.com/upload/asset-1001?signature=demo
                    expiresAt: '2026-09-28T12:10:00Z'
        '401':
          description: Authentication token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                401_upload_unauthorized:
                  value:
                    code: UNAUTHORIZED
                    message: Authentication token is missing or invalid.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '422':
          description: Unsupported type or file too large.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                422_upload_too_large:
                  value:
                    code: MEDIA_INVALID
                    message: Unsupported type or file too large.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
...
components:
  schemas:
    StandardError:
      type: object
      required:
      - code
      - message
      - requestId
      properties:
        code:
          type: string
          example: BOOKING_CONFLICT
        message:
          type: string
          example: The selected dates are already booked.
        requestId:
          type: string
          format: uuid
          example: 11111111-1111-4111-8111-111111111111
        details:
          type: object
          additionalProperties: true
          nullable: true
      example:
        code: BOOKING_CONFLICT
        message: The selected dates are already booked.
        requestId: 11111111-1111-4111-8111-111111111111
        details:
          propertyId: prop-1001
...
    UploadInit:
      type: object
      required:
      - assetId
      - uploadUrl
      - expiresAt
      properties:
        assetId:
          type: string
          example: asset-1001
        uploadUrl:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT